app/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
blinker==1.8.2
cattrs==23.2.3
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorlog==6.8.2
cryptography==43.0.0
distro==1.9.0
dnspython==2.6.1
docutils==0.21.2
dominate==2.9.1
exceptiongroup==1.2.2
fastjsonschema==2.20.0
feedgenerator==2.1.0
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Login==0.6.3
Flask-MySQL==1.5.2
flask-sock==0.7.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
jwt==1.3.1
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mypy-extensions==1.0.0
openai==1.42.0
ordered-set==4.1.0
outcome==1.3.0.post0
packaging==24.1
platformdirs==4.2.2
pluggy==1.5.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.9.0
pymongo==4.8.0
PyMySQL==1.1.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
reactpy==1.0.2
redis==5.0.8
requests==2.32.3
requests-cache==1.2.1
rich==13.8.0
selenium==4.23.1
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.32
tenacity==9.0.0
tomli==2.0.1
tqdm==4.66.5
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
Unidecode==1.3.8
url-normalize==1.4.3
urllib3==2.2.2
watchfiles==0.23.0
websocket-client==1.8.0
websockets==13.0.1
Werkzeug==3.0.3
wsproto==1.2.0
